+++ CHANGELOG.md
@@ -7,9 +7,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ## [UNRELEASED]
+### Added
+ - Added a `Calendars.resolveTimeZoneId` which maps a timezone id or alias to the current TZDB identifier for that zone.
+
 ### Fixed/Changed
  - Moved calendar available formats and interval formats to schema config.
  - Split the pluralized calendar available formats to their own schema method.
+ - Added `Calendars.timeZoneStableIds` method to return CLDR stable timezone identifiers vs `Calendars.timeZoneIds` which returns an array of valid TZDB identifiers.
+ - The `Calendars.timeZoneInfo` method now accepts a timezone identifier or alias and returns a record containing the resolved TZDB identifier and exemplar city name.
 
 ## [0.13.7]
 ### Added